+++ ng/practices-ng-forms/src/lib/form-group-fluent-builder/extensions.ts
@@ -41,9 +41,10 @@ function toFormControlOptions<T>(
     controlDef: FormControlDefinition<T> | FormControlDefinitionValueOmitted
 ): FormControlOptions {
     return {
-        validators: [],
-        asyncValidators: [],
+        validators: controlDef.validators,
+        asyncValidators: controlDef.asyncValidators,
         updateOn: controlDef.updateOn,
+        nonNullable: !controlDef.nullable,
     };
 }